v4l: subdev: increase V4L2_FRAME_DESC_ENTRY_MAX to 8
commit 908f723ba3f9c6b866ef48cc67cf8a1f5612861b from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

V4L2_FRAME_DESC_ENTRY_MAX is currently set to 4. In theory it's possible
to have an arbitrary amount of streams in a single pad, so preferably
there should be no hardcoded maximum number.

However, I believe a reasonable max is 8, which would cover a CSI-2 pad
with 4 streams of pixel data and 4 streams of metadata.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>