TI: arm64: dts: ti: k3-am62: Create overlay to enable standby LPM
commit 81d2253323be06b978a8bb31470edac1033cca6c from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Allow the A53 cores to enter standby mode in the AM62 family of devices.
Standby is defined as a CPU core entering WFI. This overlay allows the
entry of standby mode for all A53 cores by defining the standby idle
state in the device tree for each core.

In Linux, the scheduler chooses when to enter standby mode. When standby
mode is entered, the CPUIdle driver makes a call to PSCI to enter
suspend. Based on the arm,psci-suspend-param property in the device
tree, PSCI validates and suspends into the chosen power mode. In the
standby idle state, the arm,psci-suspend-param property is set to standby.
CPU standby will be entered and the CPU will enter WFI.

To prevent backportability issues, standby is enabled by an overlay
instead of the idle state being directly added in the relevant device
trees. This is because there is a bug with older versions of Trusted
Firmware-A that will crash the kernel if standby is present in the
device tree.

Signed-off-by: Kendall Willis <k-willis@ti.com>
Reviewed-by: Dhruva Gole <d-gole@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>