FROMLIST: serial: 8250: omap: Support wakeup pinctrl state on suspend
commit a35be671169fe8e38d05a51e460c098062bf2300 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

UART can be used as a wakeup source for am62 from suspend to ram states.
To enable wakeup from UART am62 requires a wakeup flag being set in the
pinctrl.

If the device is marked as wakeup enabled, select the 'wakeup' pinctrl
state on suspend and restore the default pinctrl state on resume.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
Link: https://lore.kernel.org/all/20250910-uart-daisy-chain-8250-omap-v2-2-e90d44c1a9ac@ti.com/
Signed-off-by: Kendall Willis <k-willis@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>