PENDING: media: imagination: e5010: Ensure Power On During init_device
commit 42accb71d4a289a1dc332e55512a212e2fdf5a93 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

When trying to put the board into deep sleep state, a failure can occur
resulting in each of the peripherals being woken up and restoring the
proper state they were in before deep sleep attempt began. E5010 caused
a crash during this wakeup process due to a race condition where device
was not finished turning on and the driver was making attempts to write
to regusters.

A call to device writes should ensure power of the device is on.

Reported-by: Kendall Willis <k-willis@ti.com>
Signed-off-by: Brandon Brnich <b-brnich@ti.com>
Acked-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>