PENDING: media: chips-media: wave5: Fix SError When Loading Bitstream
commit 4b539ba8631f33111d7e3bd54716a401b4452f45 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Wave5 driver uses autosuspend feaure to preserve power when device has
not been used for a certain interval. Typical processing will never see
a case when there is a large enough time between loading buffers to the
device that will cause the device to turn off. However, when utilizing
gstreamer's debug logging, there can be a case when QBUF scheduling gets
delayed resulting in an attempt to a queue a buffer to a powered off
device. Ensure power of the device is on during this call.

Signed-off-by: Jackson Lee <jackson.lee@chipsnmedia.com>
Signed-off-by: Brandon Brnich <b-brnich@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>