FROMLIST: net: ethernet: ti: am65-cpsw: Add AF_XDP zero copy for TX
commit 1e6a0d3b9a7bd8a5d540491482c10065c1416179 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add zero copy support to TX path.

Introduce xsk_pool and xsk_port_id to struct am65_cpsw_tx_chn.
This way we can quickly check if the flow is setup as XSK pool
and for which port.

If the TX channel is setup as XSK pool then get the frames from
the pool and send it to the TX channel.

Link: https://lore.kernel.org/all/20251109-am65-cpsw-xdp-zc-v2-5-858f60a09d12@kernel.org/
Signed-off-by: Roger Quadros <rogerq@kernel.org>
Signed-off-by: Meghana Malladi <m-malladi@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>