TI: crypto: ti: Increase cra_priority of DTHEv2 AES algorithms
commit f87dbe125c952ab51d5421b475c262797783dc4b from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The cra_priority value is used to decide which driver to use by default
for a particular algorithm when multiple are present. A driver with
cra_priority x is selected over all others less than x. There is no
significance of the absolute value of cra_priority.

Increase the cra_priority of DTHEv2 AES-XTS and AES-CTR algorithms over
the software implementations so that they are preferred.

Signed-off-by: T Pratham <t-pratham@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>